refactor: rename module to reflect broader scope
