+++ apps/io-session-manager/src/repositories/index.ts
@@ -6,7 +6,7 @@ import * as SpidLogsRepo from "./spid-logs";
 import * as LockedProfileRepo from "./locked-profiles";
 import * as LollipopRevokeRepo from "./lollipop-revoke-queue";
 import * as NotificationsRepo from "./notifications";
-import * as LoginEventsRepo from "./login-events";
+import * as LoginEventsRepo from "./auth-session-events";
 
 export {
   FnAppRepo,
